Setting Up British Airways Data Collection
This commit marks the initial setup for our repository, where we will be working on tasks assigned by British Airways related to data collection and analysis.

In this first commit, we have added the Jupyter Notebook provided by British Airways, which serves as our starting point for data collection. The notebook contains instructions and code snippets for collecting the required data.